add manifest to hosts with prefetch file list

+++ hooks/filesystem/useFile.ts
@@ -3,6 +3,7 @@ import { PathLike } from 'fs';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
 import {
+  fetchFileFromHost,
   FileHandle,
   FS_SKELETON_PLACEHOLDER,
   OpenFileResult,
@@ -34,37 +35,17 @@ const useFileOrFetch = (
   const [refreshSignal, triggerRefresh] = useToggle();
 
   const openFile = useCallback(async () => {
-    const deletePlaceholderResult = async (): Promise<OpenFileResult> => {
-      await fs.rm(hostQualifiedPath, { force: true });
-      return 'not found';
-    };
-
     let buffer = await fs.readFile(hostQualifiedPath).catch(() => null);
     if (!buffer) return 'not found';
 
     // current file is a placeholder from the skeleton; need to fetch actual contents
     if (!noFetch && buffer.readUint32BE() === FS_SKELETON_PLACEHOLDER) {
-      const serverFile = await fetch('hosts/' + hostQualifiedPath);
-
-      if (!serverFile.ok) return deletePlaceholderResult();
-
-      const data = await serverFile.formData();
-      const metadata = data.get('metadata');
-      const contents = data.get('contents');
-
-      if (
-        !metadata ||
-        !contents ||
-        !(typeof metadata === 'string') ||
-        !(contents instanceof File)
-      ) {
-        return deletePlaceholderResult();
+      const hostFile = await fetchFileFromHost(hostname, path);
+      if (!hostFile) {
+        await fs.rm(hostQualifiedPath, { force: true });
+        return 'not found';
       }
-      // TODO: parse metadata
-
-      const bytes = await contents.arrayBuffer();
-      buffer = Buffer.from(bytes);
-      await fs.writeFile(hostQualifiedPath, buffer);
+      buffer = hostFile;
     }
 
     // for elements taking URLs (e.g. <img>)
